Refactor cryptor superclass and its subclasses
